+++ src/KristofferStrube.Blazor.SVGEditor/SVGEditor.razor.Parameters.cs
@@ -16,6 +16,9 @@ public partial class SVGEditor
     public Action<string>? InputUpdated { get; set; }
 
     [Parameter]
+    public Action? InputRendered { get; set; }
+
+    [Parameter]
     public SelectionMode SelectionMode { get; set; } = SelectionMode.WindowSelection;
 
     [Parameter]
@@ -39,6 +42,9 @@ public partial class SVGEditor
 
     [Parameter]
     public bool SnapToInteger { get; set; } = false;
+
+    [Parameter]
+    public bool HideElements { get; set; } = false;
 
     [Parameter]
     public bool DisableContextMenu { get; set; }
@@ -77,13 +83,13 @@ public partial class SVGEditor
     public bool DisableScaleLabel { get; set; }
 
     [Parameter]
-    public List<CompleteNewShapeMenuItem> CompleteNewShapeMenuItems { get; set; } = new() {
+    public List<CompleteNewShapeMenuItem> CompleteNewShapeMenuItems { get; set; } = [
         new(typeof(CompleteWithoutCloseMenuItem), (svgEditor) => svgEditor.SelectedShapes[0] is Path),
         new(typeof(RemoveLastInstruction), (svgEditor) => svgEditor.SelectedShapes[0] is Path),
-    };
+    ];
 
     [Parameter]
-    public List<SupportedAddNewSVGElementMenuItem> AddNewSVGElementMenuItems { get; set; } = new() {
+    public List<SupportedAddNewSVGElementMenuItem> AddNewSVGElementMenuItems { get; set; } = [
         new(typeof(AddNewStopFromLinearGradientMenuItem), (svgEditor, data) => data is LinearGradient),
         new(typeof(AddNewStopFromStopMenuItem), (svgEditor, data) => data is Stop),
         new(typeof(AddNewPathMenuItem), (_, data) => data is not (LinearGradient or Stop)),
@@ -96,10 +102,10 @@ public partial class SVGEditor
         new(typeof(AddNewTextMenuItem), (_, data) => data is not (LinearGradient or Stop)),
         new(typeof(AddNewGradientMenuItem), (_, data) => data is not (LinearGradient or Stop)),
         new(typeof(AddNewAnimationMenuItem), (_, data) => data is Shape shape && !shape.IsChildElement && !shape.AnimationElements.Any(a => a.AttributeName is "fill" or "stroke" or "d")),
-    };
+    ];
 
     [Parameter]
-    public List<ActionMenuItem> ActionMenuItems { get; set; } = new() {
+    public List<ActionMenuItem> ActionMenuItems { get; set; } = [
         new(typeof(FillMenuItem), (_, data) => data is Shape shape && !shape.IsChildElement),
         new(typeof(StrokeMenuItem), (_, data) => data is Shape shape && !shape.IsChildElement),
         new(typeof(TextMenuItem), (_, data) => data is Text text && !text.IsChildElement),
@@ -112,30 +118,30 @@ public partial class SVGEditor
         new(typeof(CopyMenuItem), (svgEditor, data) => data is Shape && !svgEditor.DisableCopyElement),
         new(typeof(PasteMenuItem), (svgEditor, _) => !svgEditor.DisablePasteElement),
         new(typeof(OptimizeMenuItem), (_, data) => data is Shape shape && !shape.IsChildElement),
-    };
+    ];
 
     [Parameter]
-    public List<SupportedElement> SupportedElements { get; set; } = new() {
-        new(typeof(Rect), (IElement element) => element.TagName == "RECT"),
-        new(typeof(Circle), (IElement element) => element.TagName == "CIRCLE"),
-        new(typeof(Ellipse), (IElement element) => element.TagName == "ELLIPSE"),
-        new(typeof(Polygon), (IElement element) => element.TagName == "POLYGON"),
-        new(typeof(Polyline), (IElement element) => element.TagName == "POLYLINE"),
-        new(typeof(Line), (IElement element) => element.TagName == "LINE"),
-        new(typeof(Path), (IElement element) => element.TagName == "PATH"),
-        new(typeof(Text), (IElement element) => element.TagName == "TEXT"),
-        new(typeof(G), (IElement element) => element.TagName == "G"),
-        new(typeof(Defs), (IElement element) => element.TagName == "DEFS"),
-    };
+    public List<SupportedElement> SupportedElements { get; set; } = [
+        new(typeof(Rect), (IElement element) => element.TagName.ToUpper() == "RECT"),
+        new(typeof(Circle), (IElement element) => element.TagName.ToUpper() == "CIRCLE"),
+        new(typeof(Ellipse), (IElement element) => element.TagName.ToUpper() == "ELLIPSE"),
+        new(typeof(Polygon), (IElement element) => element.TagName.ToUpper() == "POLYGON"),
+        new(typeof(Polyline), (IElement element) => element.TagName.ToUpper() == "POLYLINE"),
+        new(typeof(Line), (IElement element) => element.TagName.ToUpper() == "LINE"),
+        new(typeof(Path), (IElement element) => element.TagName.ToUpper() == "PATH"),
+        new(typeof(Text), (IElement element) => element.TagName.ToUpper() == "TEXT"),
+        new(typeof(G), (IElement element) => element.TagName.ToUpper() == "G"),
+        new(typeof(Defs), (IElement element) => element.TagName.ToUpper() == "DEFS"),
+    ];
 
     [Parameter]
-    public List<SupportedAnimation> AnimationTypes { get; set; } = new()
-    {
+    public List<SupportedAnimation> AnimationTypes { get; set; } =
+    [
         new(typeof(AnimateFill), "fill"),
         new(typeof(AnimateStroke), "stroke"),
         new(typeof(AnimateStrokeDashoffset), "stroke-dashoffset"),
         new(typeof(AnimateD), "d"),
-    };
+    ];
 
     public void DisableAllInteractions()
     {