+++ src/KristofferStrube.Blazor.SVGEditor/SVGEditor.razor.cs
@@ -2,6 +2,7 @@
 using AngleSharp.Dom;
 using Microsoft.AspNetCore.Components;
 using Microsoft.JSInterop;
+using System.Globalization;
 using System.Reactive.Linq;
 using System.Reactive.Subjects;
 
@@ -44,15 +45,15 @@ public partial class SVGEditor : ComponentBase
 
     public List<ISVGElement> Elements { get; protected set; } = default!;
 
-    public List<Shape> SelectedShapes { get; set; } = new();
+    public List<Shape> SelectedShapes { get; set; } = [];
 
-    public Dictionary<string, ISVGElement> Definitions { get; set; } = new();
+    public Dictionary<string, ISVGElement> Definitions { get; set; } = [];
 
     public ISVGElement? EditGradient { get; set; }
 
     public Shape? FocusedShape { get; set; }
 
-    public List<Shape> MoveOverShapes { get; set; } = new();
+    public List<Shape> MoveOverShapes { get; set; } = [];
 
     public (double x, double y) MovePanner { get; set; }
 
@@ -104,11 +105,10 @@ public partial class SVGEditor : ComponentBase
         Definitions.Clear();
         ClearSelectedShapes();
 
-        IBrowsingContext context = BrowsingContext.New();
-        Document = await context.OpenAsync(req => req.Content(Input));
+        Document = await BrowsingContext.New().OpenAsync(res => res.Content($"<svg>{input}</svg>"));
 
         Elements = [];
-        foreach (IElement child in Document.GetElementsByTagName("BODY")[0].Children)
+        foreach (IElement child in Document.GetElementsByTagName("BODY")[0].Children.First().Children)
         {
             ISVGElement? sVGElement = SupportedElements.FirstOrDefault(se => se.CanHandle(child))?.ElementType is Type type
                 ? Activator.CreateInstance(type, child, this) as ISVGElement
@@ -304,5 +304,71 @@ public partial class SVGEditor : ComponentBase
             return;
         }
         Translate = (Translate.x + ((x - Translate.x) * (1 - (Scale / prevScale))), Translate.y + ((y - Translate.y) * (1 - (Scale / prevScale))));
+    }
+
+    /// <summary>
+    /// Fits all shapes in the editor to the viewport.
+    /// </summary>
+    /// <param name="delta">
+    ///     A number between <c>0</c> and <c>1</c> inclusive that defines how much of the full transformation should be made.
+    ///     <c>1</c> mean that it fits the shapes perfectly in the center.
+    ///     <c>0.5</c> means that it makes half the zoom and translation change torwards the perfect fit.
+    /// </param>
+    /// <param name="padding">How much padding there should be around the shapes to the edge of the editor. The unit is pixels in the rendered space not the space of the SVG that is edited.</param>
+    public void FitViewportToAllShapes(double delta = 1, double padding = 20)
+    {
+        FitViewportToShapes(Elements.Where(e => e is Shape).Cast<Shape>(), delta, padding);
+    }
+
+    /// <summary>
+    /// Fits the selected shapes in the editor to the viewport.
+    /// </summary>
+    /// <param name="delta">
+    ///     A number between <c>0</c> and <c>1</c> inclusive that defines how much of the full transformation should be made.
+    ///     <c>1</c> mean that it fits the shapes perfectly in the center.
+    ///     <c>0.5</c> means that it makes half the zoom and translation change torwards the perfect fit.
+    /// </param>
+    /// <param name="padding">How much padding there should be around the shapes to the edge of the editor. The unit is pixels in the rendered space not the space of the SVG that is edited.</param>
+    public void FitViewporToSelectedShapes(double delta = 1, double padding = 20)
+    {
+        FitViewportToShapes(SelectedShapes, delta, padding);
+    }
+
+    /// <summary>
+    /// Fits the <paramref name="shapes"/> to the viewport.
+    /// </summary>
+    /// <param name="shapes">The shapes to fit the viewport to.</param>
+    /// <param name="delta">
+    ///     A number between <c>0</c> and <c>1</c> inclusive that defines how much of the full transformation should be made.<br />
+    ///     <c>1</c> mean that it fits the shapes perfectly in the center.<br />
+    ///     <c>0.5</c> means that it makes half the zoom and translation change torwards the perfect fit.
+    /// </param>
+    /// <param name="padding">How much padding there should be around the shapes to the edge of the editor. The unit is pixels in the rendered space not the space of the SVG that is edited.</param>
+    public void FitViewportToShapes(IEnumerable<Shape> shapes, double delta = 1, double padding = 20)
+    {
+        if (BBox is null)
+        {
+            return;
+        }
+
+        double lowerX = double.MaxValue, lowerY = double.MaxValue;
+        double upperX = double.MinValue, upperY = double.MinValue;
+        foreach (Shape shape in shapes)
+        {
+            foreach ((double x, double y) in shape.SelectionPoints)
+            {
+                double strokeWidth = double.TryParse(shape.StrokeWidth, CultureInfo.InvariantCulture, out double result) ? result : 0;
+                lowerX = Math.Min(x - strokeWidth, lowerX);
+                upperX = Math.Max(x + strokeWidth, upperX);
+                lowerY = Math.Min(y - strokeWidth, lowerY);
+                upperY = Math.Max(y + strokeWidth, upperY);
+            }
+        }
+        double elementsWidth = upperX - lowerX;
+        double elementsHeight = upperY - lowerY;
+        double newScale = Math.Min((BBox.Width - (padding * 2)) / elementsWidth, (BBox.Height - (padding * 2)) / elementsHeight);
+        (double x, double y) newTranslate = ((BBox.Width / 2) - ((lowerX + (elementsWidth / 2)) * newScale), (BBox.Height / 2) - ((lowerY + (elementsHeight / 2)) * newScale));
+        Scale = (Scale * (1 - delta)) + (newScale * delta);
+        Translate = ((Translate.x * (1 - delta)) + (newTranslate.x * delta), (Translate.y * (1 - delta)) + (newTranslate.y * delta));
     }
 }